oneplus2: add GPS overlay
xtra server url need to be set, it is empty
in framework, without those xtra server set,
it will not download the gps almanacs,
ephemerides data from wifi or radio network. It will
need to be download from gps satellites directly
which is very slow (> 10 minutes if download from a
single satellite).

Change-Id: Iaa67aa9979a6b7f4cde40a050a574e1e7c84cdcf
Signed-off-by: Martinusbe <martinusbe@gmail.com>

Conflicts:
	overlay/frameworks/base/core/res/res/values/config.xml